Refactor: Move DestinationSet to its own module
